+++ src/main/java/com/github/jlangch/venice/impl/util/StringUtil.java
@@ -25,9 +25,12 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.StringReader;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
+
+import org.apache.commons.lang3.StringUtils;
 
 
 public class StringUtil {
@@ -130,6 +133,42 @@ public class StringUtil {
 
         return buf.toString();
     }
+
+	/**
+	 * Splits a string into columns
+	 *
+	 * @param text	a string
+	 * @param colStartPos	a list of colum start pos
+	 *
+	 * @return the splitted columns
+	 */
+	public static List<String> splitColumns(final String text, final int[] colStartPos) {
+		if (colStartPos == null || colStartPos.length == 0) {
+			throw new IllegalArgumentException("A 'colStartPos' array must not be null or empty");
+		}
+
+		final List<String> columns = new ArrayList<>();
+
+		String tmp = text;
+
+		for(int ii=colStartPos.length-1; ii>=0; ii--) {
+			int pos = colStartPos[ii];
+			if (pos <= 0) {
+				columns.add(StringUtils.trimToEmpty(tmp));
+				tmp = "";
+			}
+			else if (pos >= tmp.length()) {
+				columns.add("");
+			}
+			else {
+				columns.add(StringUtils.trimToEmpty(tmp.substring(pos)));
+				tmp = tmp.substring(0, pos);
+			}
+		}
+
+		Collections.reverse(columns);
+		return columns;
+	}
 
     /**
      * Splits a text into lines