+++ src/test/java/com/github/jlangch/venice/impl/util/StringUtilTest.java
@@ -123,4 +123,22 @@ public class StringUtilTest {
         assertEquals("3", lines.get(2));
     }
 
+
+    @Test
+    public void splitColumns() {
+        assertEquals("||", String.join("|", StringUtil.splitColumns("",     new int[] {0,1,2})));
+        assertEquals("||", String.join("|", StringUtil.splitColumns(" ",    new int[] {0,1,2})));
+        assertEquals("||", String.join("|", StringUtil.splitColumns("  ",   new int[] {0,1,2})));
+        assertEquals("||", String.join("|", StringUtil.splitColumns("   ",  new int[] {0,1,2})));
+        assertEquals("||", String.join("|", StringUtil.splitColumns("    ", new int[] {0,1,2})));
+
+        assertEquals("1|2|3", String.join("|", StringUtil.splitColumns("123", new int[] {0,1,2})));
+        assertEquals("12|34|56", String.join("|", StringUtil.splitColumns("123456", new int[] {0,2,4})));
+        assertEquals("1|2|3", String.join("|", StringUtil.splitColumns("1 2 3 ", new int[] {0,2,4})));
+
+        assertEquals("123|456|", String.join("|", StringUtil.splitColumns("123456", new int[] {0,3,6})));
+        assertEquals("123456||", String.join("|", StringUtil.splitColumns("123456", new int[] {0,6,12})));
+        assertEquals("||",       String.join("|", StringUtil.splitColumns("123456", new int[] {20,30,40})));
+    }
+
 }